unittest_requirements.txt
mock
# include the top-level requirements.txt contents below
pulpcore>=3.18.0,<3.20
ecdsa~=0.14
pyjwkest~=1.4.0
pyjwt[crypto]~=1.7.1
pytest-django
